\name{airGRdatasets-package}
\alias{airGRdatasets-package}
\alias{airGRdatasets}


\docType{package}
\encoding{UTF-8}


\title{\packageTitle{airGRdatasets}}


\description{
  \packageDescription{airGRdatasets}
  More especially, it can be used by teachers and students for hydrological modeling exercises adapted to the 'airGRteaching' package (\doi{10.32614/CRAN.package.airGRteaching}), as described in Delaigue et al. (2023) and in the 'airGRteaching' vignettes.

 The package contains the following types of datasets:
  \itemize{
    \item \code{\link{lumped_daily}}: a collection of daily hydrometeorological time series of lumped catchments, along with associated metadata.
  }
}


\source{
Delaigue, O., Guimarães, G. M., Brigode, P., Génot, B., Perrin, C., and Andréassian, V. (2024).
  CAMELS-FR dataset, V1,
  Recherche Data Gouv [data set],
  \doi{10.57745/WH7FJR}.
}


\references{
Delaigue, O., Brigode, P., Thirel, G., and Coron, L. (2023).
  airGRteaching: an open-source tool for teaching hydrological modeling with R,
  Hydrol. Earth Syst. Sci., 27, 3293–3327,
  \doi{10.5194/hess-27-3293-2023}.
\cr\cr
Delaigue, O., Guimarães, G. M., Brigode, P., Génot, B., Perrin, C., Soubeyroux, J.-M., Janet, B., Addor, N., and Andréassian, V. (2025).
  CAMELS-FR dataset: a large-sample hydroclimatic dataset for France to explore hydrological diversity and support model benchmarking,
  Earth Syst. Sci. Data, 17, 1461–1479,
  \doi{10.5194/essd-17-1461-2025}.
}


\concept{airGR}
\concept{Hydroportail}
\concept{PHyC}
\concept{SAFRAN}

