\name{airGRdatasets-package}
\alias{airGRdatasets-package}
\alias{airGRdatasets}


\docType{package}
\encoding{UTF-8}


\title{\packageTitle{airGRdatasets}}


\description{
  \packageDescription{airGRdatasets}
  More especially, it can be used by teachers and students for hydrological modeling exercises adapted to the 'airGRteaching' package as described in Delaigue et al. (submitted) and in the 'airGRteaching' vignettes.

 The package contains the following types of datasets:
  \itemize{
    \item \code{\link{lumped_daily}}: a set of lumped catchment daily hydrometeorological time series, and associated metadata.
  }
}


\source{
Delaigue, O., Brigode, P., Andréassian, V., Perrin, C., Etchevers, P., Soubeyroux, J.-M., Janet, B. and Addor, N. (2022).
CAMELS-FR: A large sample hydroclimatic dataset for France to explore hydrological diversity and support model benchmarking.
11th edition of the IAHS Scientific Assembly. International Association of Hydrological Sciences, Montpellier, 29 May - 3 Jun. 2022.
\href{https://hal.science/hal-03687235/}{hal-03687235}
}

\references{
Delaigue, O., Brigode, P., Thirel, G. and Coron, L. (submitted).
airGRteaching: an open-source tool for teaching hydrological modelling.
Hydrology and Earth System Sciences.
}


\concept{airGR}
\concept{Hydroportail}
\concept{PHyC}
\concept{SAFRAN}

